`@remotion/example`: Refactor HtmlInCanvas demo into reusable component
Splits the HtmlInCanvas demo into a reusable `<HtmlInCanvas>` component
that accepts an `applyEffect` callback, separating the WICG html-in-canvas
plumbing from the per-demo effect implementation. The wave effect is moved
into its own module, and `index.tsx` becomes a small showcase composition.

Made-with: Cursor
